Connected all the Scenes.
In Tiffany's it's Cake 5, In Macy's it's the Cupcake, In Sora's it's one of the Donuts on the hills
Look through and see if there's anything else you want to update by Tuesday.
Also think of music you guys may want to add.

Also someone should get on a start menu.

Thanks guys!